Rename the install-method helper to read as a helper, not an element
install-method-select.ts read like a custom element by the repo's
file-name-matches-element convention, but it exports a single pure function.
Rename to apply-install-method.ts (matching applyInstallMethod) so a reader
isn't hunting for an <esphome-install-method-select> that doesn't exist.
