+++ src/components/apply-install-method.ts
